==> 1_2_NatCom.2024.Analysis.240921.Rmd <==
---
title: "scBOC CD45"
author: "Siwakorn"
date: '2023-04-04'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#Library
```{r}
library(Seurat)
library(tidyverse)
library(harmony)
library(scales)
#library(pathfindR)
library(fgsea)
library(msigdbr)
library(biomaRt)
library(ggrepel)
library(ggthemes)
library(gtools)
library(DESeq2)
library("org.Mm.eg.db")
library(limma)
library(extrafont)
#font_import()
```


#==========BOC=============
#Initialization
```{r}
Batch.scRNA = c("scBRCA1","scBRCA3","scBRCA4","scBRCA5","scBRCA6", 
                paste0("scOVCA",1:5),
                paste0("scCRC",1:5))
Batch.original = c(paste0("BRCA",c(10,12,17,20,22)),
                   paste0("HGSOC",c(4,5,"5CMT",6,16)),
                   paste0("CRC",c(1,5,7,8,9)) )
ID = paste0(Batch.scRNA,"_CD45")
Condition = rep(c("CD45"),15)
Cancer = c(rep("BRCA",5),rep("OVCA",5), rep("CRC",5) )
Histo = c("Luminal","Luminal","TNBC","TNBC","TNBC",
          "HGSOC","HGSOC","HGSOC","HGSOC","HGSOC",
          "CRC","CRC","CRC","CRC","CRC")

CellRangerDir = paste0("//home/siwakorn/Human/Counts/",ID, "/filtered_feature_bc_matrix")

RNA = list()
for(i in 1:length(CellRangerDir)){
      tmp = Read10X(data.dir = CellRangerDir[i]  )
      RNA[[i]] <- CreateSeuratObject(counts = tmp, min.cells = 3, min.features = 0)
      RNA[[i]][["percent.mt"]] <- PercentageFeatureSet(RNA[[i]], pattern = "^MT-")
      RNA[[i]]$CB.original = colnames(RNA[[i]])
      RNA[[i]]$Cancer = as.character(Cancer[i])
      RNA[[i]]$Histo = as.character(Histo[i])
      RNA[[i]]$Batch.original <- as.character(Batch.original[i])
      RNA[[i]]$Batch.scRNA <- as.character(Batch.scRNA[i])
      RNA[[i]]$Condition <- as.character(Condition[i])
      RNA[[i]]$ID = paste0(as.character(RNA[[i]]$Batch.scRNA),"_",as.character(RNA[[i]]$Condition))
      RNA[[i]]$CB.new = paste0(RNA[[i]]$ID,":", colnames(RNA[[i]]) )
      RNA[[i]]$CB.new = gsub("-1","x",RNA[[i]]$CB.new)
      RNA[[i]] <- RenameCells(RNA[[i]], new.names = RNA[[i]]$CB.new )
}

print("part1")
BOC = merge(x = RNA[[1]],
            y = RNA[2:length(RNA)],
            merge.data = TRUE)
RNA = 0
DefaultAssay(BOC) = "RNA"

print("part2")
BOC = NormalizeData(BOC) %>% FindVariableFeatures() %>% ScaleData() %>% RunPCA(verbose = FALSE)
BOC = RunHarmony(BOC, group.by.vars = "ID")

saveRDS(BOC, file = "//home/siwakorn/Human/RDS/BOC.CD45.230712.rds") 

BOC = subset(BOC, percent.mt < 10)
BOC = NormalizeData(BOC) %>% FindVariableFeatures() %>% ScaleData() %>% RunPCA(verbose = FALSE)
BOC = RunHarmony(BOC, group.by.vars = "ID")
BOC <- RunUMAP(BOC, reduction = "harmony", dims = 1:30, min.dist = 1, spread =1 )
BOC <- FindNeighbors(BOC, reduction = "harmony", dims = 1:30) %>% FindClusters(resolution = 0.3)
saveRDS(BOC, file = "//home/siwakorn/Human/RDS/BOC_preSubmitted.QC10.230712.rds")
```


#Classification
```{r}
BOC <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC_preSubmitted.QC10.230712.rds")
BOC = RenameIdents(BOC, 
                   '0' = "TNK",
                   '3' = "TNK",
                   '4' = "TNK",
                   '6' = "TNK",
                   '9' = "TNK",
                   '14' = "TNK",
                   '17' = "TNK",
                   '18' = "TNK",
                   '2' = "Myeloid",
                   '8' = "Myeloid",
                   '15' = "Myeloid",
                   '16' = "Myeloid",
                   '5' = "B_cell",
                   '1' = "B_cell",
                   '7' = "Dividing",
                   '12' = "Dividing",
                   '13' = "Dividing",
                   '10' = "Minor",
                   '11' = "Minor"
                   )
BOC$Identity.Lineage = Idents(BOC)


#-----------------------------------------------------------------------------------------
##TNK
TNK = subset(BOC,idents = "TNK")
TNK = Clustering1(TNK)
TNK = RenameIdents(TNK,
                   '5' = "NK",
                   '6' = "NK",
                   '0' = "T_CD8",
                   '4' = "T_CD8",
                   '7' = "T_CD8",
                   '1' = "T_CD4",
                   '2' = "T_CD4",
                   '3' = "Treg",
                   '8' = "LQ",
                   '9' = "LQ")
TNK$Identity.Celltype = Idents(TNK)

NK = subset(TNK, idents = "NK")
NK = Clustering1(NK)
NK = RenameIdents(NK,
                   '0' = "T_CD8",
                   '1' = "NK",
                   '2' = "T_CD8",
                   '3' = "NK",
                   '4' = "NK",
                   '5' = "NK",
                   '6' = "NK",
                   '7' = "NK",
                   '8' = "NK",
                   '9' = "NK")
NK$Identity.Celltype = Idents(NK)

CD8 = subset(TNK, idents = "T_CD8")
CD8 = Clustering1(CD8)
CD8 = RenameIdents(CD8,
                   '0' = "T_CD8",
                   '1' = "T_CD8",
                   '3' = "T_CD8",
                   '6' = "T_CD8")
CD8$Identity.Celltype = Idents(CD8)
C2 = subset(CD8, idents = 2)
C2 = Clustering1(C2)
C2 = RenameIdents(C2,
                  '0' = "T_CD8",
                  '2' = "T_CD8",
                  '1' = "T_CD4",
                  '4' = "T_CD4",
                  '3' = "LQ",
                  '5' = "LQ")
C2$Identity.Celltype = Idents(C2)

C4 = subset(CD8, idents = 4)
C4 = Clustering1(C4,res =1)
C4 = RenameIdents(C4,
                   '0' = "T_CD4",
                   '1' = "T_CD8",
                   '2' = "T_CD8",
                   '3' = "T_CD4",
                   '4' = "T_CD4",
                   '5' = "T_CD8",
                   '6' = "T_CD8",
                   '7' = "T_CD8",
                   '8' = "T_CD8",
                   '9' = "T_CD4",
                   '10' = "T_CD4"
                  )
C4$Identity.Celltype = Idents(C4)

C5 = subset(CD8, idents = 5)
C5 = Clustering1(C5)
C5 = RenameIdents(C5, 
                  '0' = "T_CD4",
                  '1' = "T_CD8",
                  '2' = "T_CD4",
                  '3' = "LQ",
                  '4' = "T_CD4")
C5$Identity.Celltype = Idents(C5)

tmp1 = CD8[[]] %>% dplyr::select(Identity.Celltype) %>% filter(Identity.Celltype == "T_CD8")
tmp2 = C2[[]] %>% dplyr::select(Identity.Celltype) 
tmp3 = C4[[]] %>% dplyr::select(Identity.Celltype) 
tmp4 = C5[[]] %>% dplyr::select(Identity.Celltype) 

tmp = rbind(tmp1,tmp2,tmp3,tmp4)
tmp1 = tmp$Identity.Celltype
names(tmp1) = rownames(tmp)
CD8 = AddMetaData(CD8,tmp1, "Identity.Celltype")

CD4 = subset(TNK, idents = "T_CD4")
CD4 = Clustering1(CD4)
CD4 = RenameIdents(CD4,
                   '0' = "T_CD4",
                   '1' = "T_CD4",
                   '2' = "T_CD4",
                   '3' = "T_CD4",
                   '4' = "T_CD8",
                   '5' = "T_CD4",
                   '6' = "T_CD8"
                   )
CD4$Identity.Celltype = Idents(CD4)

Treg= subset(TNK, idents = "Treg")
Treg = Clustering1(Treg)
Treg = RenameIdents(Treg,
                    '3' = "T_CD8",
                    '0' = "T_CD4_FOXP3",
                    '1' = "T_CD4_FOXP3",
                    '2' = "T_CD4_FOXP3",
                    '4' = "T_CD4_FOXP3",
                    '5' = "T_CD4_FOXP3",
                    '6' = "T_CD4_FOXP3")
Treg$Identity.Celltype = Idents(Treg)


tmp1 = TNK[[]] %>% dplyr::select(Identity.Celltype) %>% filter(Identity.Celltype == "LQ")
tmp2 = NK[[]] %>% dplyr::select(Identity.Celltype) 
tmp3 = CD8[[]] %>% dplyr::select(Identity.Celltype) 
tmp4 = CD4[[]] %>% dplyr::select(Identity.Celltype) 
tmp5 = Treg[[]] %>% dplyr::select(Identity.Celltype) 

tmp = rbind(tmp1,tmp2,tmp3,tmp4,tmp5)
tmp1 = tmp$Identity.Celltype
names(tmp1) = rownames(tmp)
TNK = AddMetaData(TNK,tmp1, "Identity.Celltype")
Idents(TNK) = TNK$Identity.Celltype
DimPlot(TNK)
#-----------------------------------------------------------------------------------------
##Myeloid
M = subset(BOC, idents = "Myeloid")
M = Clustering1(M)
DimPlot(M,label = T)
Idents(M) = M$seurat_clusters
M = RenameIdents(M,
                 '0' = "TIM",
                 '2' = "TIM",
                 '3' = "TIM",
                 '6' = "TIM",
                 '8' = "TIM",
                 '11' = "TIM",
                 '5' = "TIM_Dividing",
                 '1' = "TAN-PMN",
                 '4' = "cDC2",
                 '10' = "cDC1",
                 '9' = "mregDC",
                 '7' = "pDC"
                 )
M$Identity.Celltype = Idents(M)
TIM = subset(M, idents = "TIM")
TIM = Clustering1(TIM)
TIM = RenameIdents(TIM,
                   '0' = "TIM_VCAN",
                   '1' = "TIM_C1QA",
                   '2' = "C2",
                   '3' = "TIM_VCAN",
                   '4' = "TIM_C1QA",
                   '5' = "Minor",
                   '6' = "Minor",
                   '7' = "Minor",
                   '8' = "Minor")
TIM$Identity.Celltype = Idents(TIM)

C2 = subset(TIM, idents = "C2")
C2 = Clustering1(C2)
C2 = RenameIdents(C2, 
                  '1' = "TIM_VCAN",
                  '2' = "TIM_C1QA",
                  '3' = "LQ",
                  '5' = "TIM_C1QA",
                  '6' = "TIM_C1QA",
                  '0' = "LQ",
                  '4' = "LQ",
                  '7' = "LQ")
C2$Identity.Celltype = Idents(C2)

C5 = subset(M,idents = "TIM_Dividing")
C5 = Clustering1(C5)

tmp1 = M[[]] %>% dplyr::select(Identity.Celltype) %>% filter(!Identity.Celltype == "TIM")
tmp2 = TIM[[]]  %>% dplyr::select(Identity.Celltype) %>% filter(!Identity.Celltype == "C2")
tmp3 = C2[[]]  %>% dplyr::select(Identity.Celltype)
tmp = rbind(tmp1,tmp2,tmp3)
tmp1 = tmp$Identity.Celltype
names(tmp1) = rownames(tmp)
M = AddMetaData(M, tmp1, "Identity.Celltype")
#-----------------------------------------------------------------------------------------
##Div
Div = subset(BOC, idents = "Dividing")
Div = Clustering1(Div)
Div$seurat_clusters = Idents(Div)
Div = RenameIdents(Div,
                   '0' = "T_CD8_Dividing",
                   '1' = "T_CD4_Dividing",
                   '2' = "T_CD4_FOXP3_Dividing",
                   '5' = "T_GD_Dividing",
                   '2' = "B_cell_Dividing",
                   '3' = "Plasma_cell_Dividing",
                   '7' = "Plasma_cell_Dividing",
                   '6' = "LQ")

Idents(Div) = Div$seurat_clusters
Div = RenameIdents(Div,
                   '0' = "T_CD8",
                   '1' = "T_CD4",
                   '4' = "T_CD4_FOXP3",
                   '5' = "NK",
                   '2' = "B_cell",
                   '3' = "B_cell",
                   '7' = "B_cell",
                   '6' = "LQ")
Div$Identity.Celltype = Idents(Div)
DimPlot(Div,label=T)


#-----------------------------------------------------------------------------------------
#Embed
BOC$Identity.Celltype = BOC$Identity.Lineage
tmp1 = BOC[[]] %>% dplyr::select(Identity.Celltype) %>% filter(Identity.Celltype %in% c("B_cell","Minor") )
tmp2 = TNK[[]] %>% dplyr::select(Identity.Celltype) 
tmp3 = M[[]] %>% dplyr::select(Identity.Celltype) 
tmp4 = Div[[]] %>% dplyr::select(Identity.Celltype) 
tmp = rbind(tmp1,tmp2,tmp3,tmp4)
tmp1 = tmp$Identity.Celltype
names(tmp1)= rownames(tmp)

BOC = AddMetaData(BOC, tmp1, "Identity.Celltype")
write.csv(sort(unique(BOC$Identity.Celltype)), "BOC.Identity.Celltype.230406.csv")
tmp =read.csv("BOC.Identity.Celltype.230406.modified.csv")$x
BOC$Identity.Celltype = factor(BOC$Identity.Celltype, levels= tmp)
BOC$Identity.Celltype1 = BOC$Identity.Celltype
Idents(BOC) = BOC$Identity.Celltype

#NK
NK = subset(BOC, idents = "NK")
NK = Clustering1(NK)
NK = RenameIdents(NK,
                  '0' = "NK",
                   '1' = "NK",
                   '2' = "NK",
                   '3' = "NK",
                   '4' = "NK",
                   '5' = "NK",
                   '6' = "NK",
                   '7' = "NK",
                   '9' = "NK_Dividing",
                   '10' = "NK",
                  '8' = "LQ")
NK$Identity.Celltype2 = Idents(NK)
NK$Identity.Subtype1 = Idents(NK)
NK$Identity.Subtype2 = Idents(NK)
DimPlot(NK,label=T)

#-----------------------------------------------------------------------------------------
#CD8
CD8 = subset(BOC, idents = "T_CD8")
CD8 = Clustering1(CD8)
Idents(CD8) = CD8$seurat_clusters
CD8 = RenameIdents(CD8, 
                   '0' = "T_CD8",
                   '1' = "T_CD8",
                   '2' = "T_CD8",
                   '3' = "T_CD8",
                   '4' = "T_CD8_Dividing",
                   '5' = "T_CD8",
                   '6' = "T_CD8",
                   '7' = "T_CD8",
                   '8' = "T_CD8",
                   '9' = "T_CD8",
                   '10' = "T_CD8" #KLRD1
                   )
CD8$Identity.Celltype2 = Idents(CD8)
Idents(CD8) = CD8$seurat_clusters
CD8 = RenameIdents(CD8, 
                    '3' = "T_CD8_TCF7",
                   '0' = "T_CD8_GZMK",
                   '2' = "T_CD8_GZMK",
                   '1' = "T_CD8_HAVCR2",
                   '5' = "T_CD8_GNLY",
                   '10' = "T_CD8_GNLY", #KLRD1
                   '6' = "T_CD8_ISG",
                   '8' = "T_CD8_RORA",
                   '9' = "T_CD8_TNFRSF4",
                   '7' = "T_CD8_FOXP3",
                   '4' = "T_CD8_Dividing"
                   )
CD8$Identity.Subtype1 = Idents(CD8)
Idents(CD8) = CD8$seurat_clusters
CD8 = RenameIdents(CD8, 
                    '3' = "T_CD8_TCF7",
                   '0' = "T_CD8_GZMK_s1",
                   '2' = "T_CD8_GZMK_s2",
                   '1' = "T_CD8_HAVCR2",
                   '5' = "T_CD8_GNLY_s1",
                   '10' = "T_CD8_GNLY_s2", #KLRD1
                   '6' = "T_CD8_ISG",
                   '8' = "T_CD8_RORA",
                   '9' = "T_CD8_TNFRSF4",
                   '7' = "T_CD8_FOXP3",
                   '4' = "T_CD8_Dividing"
                   )
CD8$Identity.Subtype2 = Idents(CD8)
DimPlot(CD8, label= T)
#-----------------------------------------------------------------------------------------
#CD4
CD4 = subset(BOC, idents = "T_CD4")
CD4 = Clustering1(CD4)

Idents(CD4) = CD4$seurat_clusters
CD4 = RenameIdents(CD4,
                   '0' = "T_CD4",
                   '1' = "T_CD4",
                   '2' = "T_CD4",
                   '3' = "T_CD4",
                   '4' = "T_CD4",
                   '5' = "T_CD4",
                   '8' = "T_CD4",
                   '7' = "T_CD4", #ISG
                   '6' = "T_CD4_Dividing",
                   '9' = "LQ")
CD4$Identity.Celltype2 = Idents(CD4)

Idents(CD4) = CD4$seurat_clusters
CD4 = RenameIdents(CD4,
                   '0' = "T_CD4_KLF2",
                   '1' = "T_CD4_TOX",
                   '2' = "T_CD4_KLF2",
                   '3' = "T_CD4_TOX",
                   '4' = "T_CD4_KLF2",
                   '5' = "T_CD4_TOX",
                   '8' = "T_CD4_TOX",
                   '7' = "T_CD4_ISG", #ISG
                   '6' = "T_CD4_Dividing",
                   '9' = "LQ")
CD4$Identity.Subtype1 = Idents(CD4)

Idents(CD4) = CD4$seurat_clusters
CD4 = RenameIdents(CD4,
                   '0' = "T_CD4_KLF2_s1",
                   '2' = "T_CD4_KLF2_s2",
                   '4' = "T_CD4_KLF2_s3",
                   '1' = "T_CD4_TOX_s1",
                   '3' = "T_CD4_TOX_s2",
                   '5' = "T_CD4_TOX_s3",
                   '8' = "T_CD4_TOX_s4",
                   '7' = "T_CD4_ISG", #ISG
                   '6' = "T_CD4_Dividing",
                   '9' = "LQ")
CD4$Identity.Subtype2 = Idents(CD4)
#-----------------------------------------------------------------------------------------
#Treg
Treg = subset(BOC,idents = "T_CD4_FOXP3")
Treg = Clustering1(Treg)
Idents(Treg)= Treg$seurat_clusters
Treg = RenameIdents(Treg, 
                   '0' = "T_CD4_FOXP3",
                   '1' = "T_CD4_FOXP3",
                   '2' = "T_CD4_FOXP3",
                   '3' = "T_CD4_FOXP3",
                   '4' = "T_CD4_FOXP3",
                   '5' = "T_CD4_FOXP3",
                   '7' = "T_CD4_FOXP3",
                   '6' = "T_CD4_FOXP3_Dividing")
Treg$Identity.Celltype2 = Idents(Treg)
Idents(Treg)= Treg$seurat_clusters
Treg = RenameIdents(Treg, 
                   '0' = "T_CD4_FOXP3_KLF2",
                   '1' = "T_CD4_FOXP3_TNFRSF9",
                   '2' = "T_CD4_FOXP3_TNFRSF9",
                   '3' = "T_CD4_FOXP3_TNFRSF9",
                   '4' = "T_CD4_FOXP3_TNFRSF9",
                   '5' = "T_CD4_FOXP3_TNFRSF9",
                   '7' = "T_CD4_FOXP3_TNFRSF9",
                   '6' = "T_CD4_FOXP3_Dividing")
Treg$Identity.Subtype1 = Idents(Treg)
Idents(Treg)= Treg$seurat_clusters
Treg = RenameIdents(Treg, 
                   '0' = "T_CD4_FOXP3_KLF2",
                   '1' = "T_CD4_FOXP3_TNFRSF9_s1",
                   '2' = "T_CD4_FOXP3_TNFRSF9_s2",
                   '3' = "T_CD4_FOXP3_TNFRSF9_s3",
                   '4' = "T_CD4_FOXP3_TNFRSF9_s4",
                   '5' = "T_CD4_FOXP3_TNFRSF9_s5",
                   '7' = "T_CD4_FOXP3_TNFRSF9_s6",
                   '6' = "T_CD4_FOXP3_Dividing")
Treg$Identity.Subtype2 = Idents(Treg)
DimPlot(Treg,label=T)
#-----------------------------------------------------------------------------------------
#TIM
TIM = subset(BOC, idents = c("TIM_VCAN","TIM_C1QA","TIM_Dividing","TAN-PMN"))
TIM = Clustering1(TIM,res = 1)
DimPlot(TIM,label=T)
Idents(TIM) = TIM$seurat_clusters
TIM = RenameIdents(TIM,
                   '0' = "TAN-PMN",
                   '6' = "TAN-PMN",
                   '15' = "TAN-PMN",
                   '10' = "TAN-PMN",
                   '1' = "TIM",
                   '7' = "TIM",
                   '9' = "TIM",
                   '18' = "TIM",
                   '13' = "TIM",
                   '14' = "TIM",
                   '8' = "TIM_Dividing",
                   '17' = "TIM_Dividing",
                   '2' = "TIM",
                   '3' = "TIM",
                   '4' = "TIM",
                   '11' = "TIM",
                   '12' = "TIM",
                   '16' = "TIM",
                   '5' = "LQ"
                   )
TIM$Identity.Celltype2 = Idents(TIM)
DimPlot(TIM)
Idents(TIM) = TIM$seurat_clusters
TIM = RenameIdents(TIM,
                   '0' = "TAN-PMN",
                   '6' = "TAN-PMN",
                   '15' = "TAN-PMN",
                   '10' = "TAN-PMN",
                   '1' = "TIM_C1QA",
                   '7' = "TIM_C1QA",
                   '9' = "TIM_C1QA",
                   '18' = "TIM_C1QA",
                   '13' = "TIM_C1QA",
                   '14' = "TIM_C1QA",
                   '8' = "TIM_C1QA_Dividing",
                   '17' = "TIM_C1QA_Dividing",
                   '2' = "TIM_VCAN",
                   '3' = "TIM_VCAN",
                   '4' = "TIM_VCAN",
                   '11' = "TIM_VCAN",
                   '12' = "TIM_VCAN",
                   '16' = "TIM_VCAN",
                   '5' = "LQ"
                   )
TIM$Identity.Subtype1 = Idents(TIM)
Idents(TIM) = TIM$seurat_clusters
TIM = RenameIdents(TIM,
                   '0' = "TAN-PMN_s1",
                   '6' = "TAN-PMN_s2",
                   '15' = "TAN-PMN_s3",
                   '10' = "TAN-PMN_s4",
                   '1' = "TIM_C1QA_s1",
                   '7' = "TIM_C1QA_s2",
                   '9' = "TIM_C1QA_s3",
                   '18' = "TIM_C1QA_s4",
                   '13' = "TIM_C1QA_s5",
                   '14' = "TIM_C1QA_s6",
                   '8' = "TIM_C1QA_Dividing_s1",
                   '17' = "TIM_C1QA_Dividing_s2",
                   '2' = "TIM_VCAN_s1",
                   '3' = "TIM_VCAN_s2",
                   '4' = "TIM_VCAN_s3",
                   '11' = "TIM_VCAN_s4",
                   '12' = "TIM_VCAN_s5",
                   '16' = "TIM_VCAN_s6",
                   '5' = "LQ"
                   )
TIM$Identity.Subtype2 = Idents(TIM)

#-----------------------------------------------------------------------------------------
#B Cell
B = subset(BOC, idents = "B_cell")
B = Clustering1(B)
Idents(B) = B$seurat_clusters
B = RenameIdents(B,
                 '2' = "B_cell",
                 '4' = "B_cell",
                 '0' = "Plasma_cell",
                 '1' = "Plasma_cell",
                 '3' = "Plasma_cell",
                 '5' = "Plasma_cell",
                 '7' = "Plasma_cell",
                 '6' = "B-Plasma_cell_Dividing")
B$Identity.Celltype2 = Idents(B)
B$Identity.Subtype1 = Idents(B)
Idents(B) = B$seurat_clusters
B = RenameIdents(B,
                 '2' = "B_cell_s1",
                 '4' = "B_cell_s2",
                 '0' = "Plasma_cell_s1",
                 '1' = "Plasma_cell_s2",
                 '3' = "Plasma_cell_s3",
                 '5' = "Plasma_cell_s4",
                 '7' = "Plasma_cell_s5",
                 '6' = "B-Plasma_cell_Dividing")
B$Identity.Subtype2 = Idents(B)

#Embed
Idents(BOC)=BOC$Identity.Celltype
DimPlot(BOC)
tmp1 = BOC[[]] %>% dplyr::select(Identity.Celltype) %>% filter(Identity.Celltype %in% c("cDC1","cDC2","mregDC","pDC","Minor","LQ"))
tmp1$Identity.Celltype2 = tmp1$Identity.Celltype
tmp1$Identity.Subtype1 = tmp1$Identity.Celltype
tmp1$Identity.Subtype2 = tmp1$Identity.Celltype
tmp2 = NK[[]]  %>% dplyr::select(Identity.Celltype,Identity.Celltype2,Identity.Subtype1,Identity.Subtype2)
tmp3 = CD8[[]] %>% dplyr::select(Identity.Celltype,Identity.Celltype2,Identity.Subtype1,Identity.Subtype2)
tmp4 = CD4[[]] %>% dplyr::select(Identity.Celltype,Identity.Celltype2,Identity.Subtype1,Identity.Subtype2)
tmp5 = Treg[[]] %>% dplyr::select(Identity.Celltype,Identity.Celltype2,Identity.Subtype1,Identity.Subtype2)
tmp6 = TIM[[]] %>% dplyr::select(Identity.Celltype,Identity.Celltype2,Identity.Subtype1,Identity.Subtype2)
tmp7 = B[[]] %>% dplyr::select(Identity.Celltype,Identity.Celltype2,Identity.Subtype1,Identity.Subtype2)
tmp = rbind(tmp1,tmp2,tmp3,tmp4,tmp5,tmp6,tmp7)

tmp1 = tmp$Identity.Celltype2
names(tmp1) = rownames(tmp)
BOC = AddMetaData(BOC,tmp1,"Identity.Celltype2")

tmp1 = tmp$Identity.Subtype1
names(tmp1) = rownames(tmp)
BOC = AddMetaData(BOC,tmp1,"Identity.Subtype1")

tmp1 = tmp$Identity.Subtype2
names(tmp1) = rownames(tmp)
BOC = AddMetaData(BOC,tmp1,"Identity.Subtype2")

write.csv(sort(unique(BOC$Identity.Celltype2)),"BOC.CD45.Identity.Celltype2.230712.csv")
write.csv(sort(unique(BOC$Identity.Subtype1)),"BOC.CD45.Identity.Subtype1.230712.csv")
write.csv(sort(unique(BOC$Identity.Subtype2)),"BOC.CD45.Identity.Subtype2.230712.csv")

tmp = c("NK","NK_Dividing","T_CD8","T_CD8_Dividing","T_CD4","T_CD4_Dividing","T_CD4_FOXP3","T_CD4_FOXP3_Dividing","TAN-PMN","TIM","TIM_Dividing","cDC1","cDC2","mregDC","pDC","B_cell","Plasma_cell","B-Plasma_cell_Dividing","Minor","LQ")
BOC$Identity.Celltype2 = factor(BOC$Identity.Celltype2, levels = tmp)

tmp = c("NK","NK_Dividing","T_CD8_TCF7","T_CD8_GZMK","T_CD8_HAVCR2","T_CD8_GNLY","T_CD8_ISG","T_CD8_RORA","T_CD8_TNFRSF4","T_CD8_FOXP3","T_CD8_Dividing","T_CD4_KLF2","T_CD4_TOX","T_CD4_ISG","T_CD4_Dividing","T_CD4_FOXP3_KLF2","T_CD4_FOXP3_TNFRSF9","T_CD4_FOXP3_Dividing","TAN-PMN","TIM_VCAN","TIM_C1QA","TIM_C1QA_Dividing","cDC1","cDC2","mregDC","pDC","B_cell","Plasma_cell","B-Plasma_cell_Dividing","Minor","LQ")
BOC$Identity.Subtype1 = factor(BOC$Identity.Subtype1, levels = tmp)

tmp = c("NK","NK_Dividing","T_CD8_TCF7","T_CD8_GZMK_s1","T_CD8_GZMK_s2","T_CD8_HAVCR2","T_CD8_GNLY_s1","T_CD8_GNLY_s2","T_CD8_ISG","T_CD8_RORA","T_CD8_TNFRSF4","T_CD8_FOXP3","T_CD8_Dividing","T_CD4_KLF2_s1","T_CD4_KLF2_s2","T_CD4_KLF2_s3","T_CD4_TOX_s1","T_CD4_TOX_s2","T_CD4_TOX_s3","T_CD4_TOX_s4","T_CD4_ISG","T_CD4_Dividing","T_CD4_FOXP3_KLF2","T_CD4_FOXP3_TNFRSF9_s1","T_CD4_FOXP3_TNFRSF9_s2","T_CD4_FOXP3_TNFRSF9_s3","T_CD4_FOXP3_TNFRSF9_s4","T_CD4_FOXP3_TNFRSF9_s5","T_CD4_FOXP3_TNFRSF9_s6","T_CD4_FOXP3_Dividing","TAN-PMN_s1","TAN-PMN_s2","TAN-PMN_s3","TAN-PMN_s4","TIM_C1QA_s1","TIM_C1QA_s2","TIM_C1QA_s3","TIM_C1QA_s4","TIM_C1QA_s5","TIM_C1QA_s6","TIM_C1QA_Dividing_s1","TIM_C1QA_Dividing_s2","TIM_VCAN_s1","TIM_VCAN_s2","TIM_VCAN_s3","TIM_VCAN_s4","TIM_VCAN_s5","TIM_VCAN_s6","cDC1","cDC2","mregDC","pDC","B_cell_s1","B_cell_s2","Plasma_cell_s1","Plasma_cell_s2","Plasma_cell_s3","Plasma_cell_s4","Plasma_cell_s5","B-Plasma_cell_Dividing","Minor","LQ")
BOC$Identity.Subtype2 = factor(BOC$Identity.Subtype2, levels = tmp)

Idents(BOC) = BOC$Identity.Celltype2
BOC <- RenameIdents(BOC,
                    'NK' = "NK",
                    'NK_Dividing' = "NK",
                    'T_CD8' = "CD8 T cell",
                    'T_CD8_Dividing' = "CD8 T cell",
                    'T_CD4' = "CD4 T cell",
                    'T_CD4_Dividing' = "CD4 T cell",
                    'T_CD4_FOXP3' = "regulatory T cell",
                    'T_CD4_FOXP3_Dividing' = "regulatory T cell",
                    'TAN-PMN' = "TAN-PMN",
                    'TIM' = "TIM",
                    'TIM_Dividing' = "TIM",
                    'cDC1' = "DC",
                    'cDC2' = "DC",
                    'mregDC' = "DC",
                    'pDC' = "DC",
                    'B_cell' = "B cell",
                    'B-Plasma_cell_Dividing' = "B cell",
                    'Plasma_cell' = "Plasma cell")
BOC$Identity.Celltype3 = Idents(BOC)

Idents(BOC) = BOC$Identity.Celltype2
BOC <- RenameIdents(BOC,
                    'NK' = "NK",
                    'NK_Dividing' = "NK",
                    'T_CD8' = "T_CD8",
                    'T_CD8_Dividing' = "T_CD8",
                    'T_CD4' = "T_CD4",
                    'T_CD4_Dividing' = "T_CD4",
                    'T_CD4_FOXP3' = "T_CD4_FOXP3",
                    'T_CD4_FOXP3_Dividing' = "T_CD4_FOXP3",
                    'TAN-PMN' = "TIM",
                    'TIM' = "TIM",
                    'TIM_Dividing' = "TIM",
                    'cDC1' = "DC",
                    'cDC2' = "DC",
                    'mregDC' = "DC",
                    'pDC' = "DC",
                    'B_cell' = "B cell",
                    'B-Plasma_cell_Dividing' = "B cell",
                    'Plasma_cell' = "Plasma cell")
BOC$Identity.Celltype = Idents(BOC)

Idents(BOC) = BOC$Identity.Celltype2
BOC <- RenameIdents(BOC,
                    'NK' = "NK",
                    'NK_Dividing' = "NK",
                    'T_CD8' = "CD8 T cell",
                    'T_CD8_Dividing' = "CD8 T cell",
                    'T_CD4' = "CD4 T cell",
                    'T_CD4_Dividing' = "CD4 T cell",
                    'T_CD4_FOXP3' = "regulatory T cell",
                    'T_CD4_FOXP3_Dividing' = "regulatory T cell",
                    'TAN-PMN' = "TIM",
                    'TIM' = "TIM",
                    'TIM_Dividing' = "TIM",
                    'cDC1' = "DC",
                    'cDC2' = "DC",
                    'mregDC' = "DC",
                    'pDC' = "DC",
                    'B_cell' = "B cell",
                    'B-Plasma_cell_Dividing' = "B cell",
                    'Plasma_cell' = "Plasma cell")
BOC$Identity.CelltypeR = Idents(BOC)

Idents(BOC) = BOC$Identity.Celltype2
BOC <- RenameIdents(BOC,
                    'NK' = "NK",
                    'T_CD8' = "T_CD8",
                    'T_CD4' = "T_CD4",
                    'T_CD4_FOXP3' = "T_CD4_FOXP3",
                    'TAN-PMN' = "TIM",
                    'TIM' = "TIM",
                    'cDC1' = "DC",
                    'cDC2' = "DC",
                    'mregDC' = "DC",
                    'pDC' = "DC",
                    'B_cell' = "B cell",
                    'Plasma_cell' = "Plasma cell",
                    'NK_Dividing' = "Dividing",
                    'T_CD8_Dividing' = "Dividing",
                    'T_CD4_Dividing' = "Dividing",
                    'T_CD4_FOXP3_Dividing' = "Dividing",
                    'TIM_Dividing' = "Dividing",
                    'B-Plasma_cell_Dividing' = "Dividing")
BOC$Identity.Celltype4 = Idents(BOC)

Idents(BOC) = BOC$Identity.Celltype2
BOC <- RenameIdents(BOC,
                    'NK' = "NK",
                    'T_CD8' = "CD8 T cell",
                    'T_CD4' = "CD4 T cell",
                    'T_CD4_FOXP3' = "regulatory T cell",
                    'TAN-PMN' = "TIM",
                    'TIM' = "TIM",
                    'cDC1' = "DC",
                    'cDC2' = "DC",
                    'mregDC' = "DC",
                    'pDC' = "DC",
                    'B_cell' = "B cell",
                    'Plasma_cell' = "Plasma cell",
                    'NK_Dividing' = "Dividing",
                    'T_CD8_Dividing' = "Dividing",
                    'T_CD4_Dividing' = "Dividing",
                    'T_CD4_FOXP3_Dividing' = "Dividing",
                    'TIM_Dividing' = "Dividing",
                    'B-Plasma_cell_Dividing' = "Dividing"
                    )
BOC$Identity.Celltype4R = Idents(BOC)
DimPlot(BOC)
saveRDS(BOC, "BOC_preSubmitted.230713.rds")
```

#Exclude Minor-LQ
```{r}
tmp = levels(Idents(BOC)) %>% setdiff(c("Minor","LQ"))
BOC = subset(BOC, idents = tmp)
BOC = RunUMAP(BOC, reduction = "harmony", dims = 1:30, min.dist = 1, spread = 1 )
saveRDS(BOC, "BOC_preSubmitted.QC_Cluster.230713.rds")
```


#PTGER4 Grouping
```{r}
tmp = PTGER4.Group(BOC,Identity = "Identity.Celltype",Batch = "Batch.scRNA" )
BOC = AddMetaData(BOC,tmp,"PTGER4.Group.Celltype")
BOC$PTGER4.Group.Celltype = factor(BOC$PTGER4.Group.Celltype, levels = c("Undetected","Low","Intermediate","High"))

tmp = PTGER4.Group(BOC,Identity = "Identity.Celltype2",Batch = "Batch.scRNA" )
BOC = AddMetaData(BOC,tmp,"PTGER4.Group.Celltype2")
BOC$PTGER4.Group.Celltype2 = factor(BOC$PTGER4.Group.Celltype2, levels = c("Undetected","Low","Intermediate","High"))


tmp = PTGER4.Group(BOC,Identity = "Identity.Celltype4",Batch = "Batch.scRNA" )
BOC = AddMetaData(BOC,tmp,"PTGER4.Group.Celltype4")
BOC$PTGER4.Group.Celltype4 = factor(BOC$PTGER4.Group.Celltype4, levels = c("Undetected","Low","Intermediate","High"))

tmp = PTGER4.Group(BOC,Identity = "Identity.Subtype1",Batch = "Batch.scRNA" )
BOC = AddMetaData(BOC,tmp,"PTGER4.Group.Subtype1")
BOC$PTGER4.Group.Subtype1 = factor(BOC$PTGER4.Group.Subtype1, levels = c("Undetected","Low","Intermediate","High"))
saveRDS(BOC,"BOC.PTGER4Group.230713.rds")
#BOC = readRDS("BOC.PTGER4Group.230713.rds")
```

#PTGER2 Grouping
```{r}
BOC <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.PTGER4Group.230713.rds")
tmp = PTGER2.Group(BOC,Identity = "Identity.Celltype",Batch = "Batch.scRNA" )
BOC = AddMetaData(BOC,tmp,"PTGER2.Group.Celltype")
BOC$PTGER2.Group.Celltype = factor(BOC$PTGER2.Group.Celltype, levels = c("Undetected","Low","High"))
saveRDS(BOC, "~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.PTGER2Group.240110.rds" )
```

#PTGER4-PTGER2 Group
```{r}
BOC$PTGER4_PTGER2 = paste0(BOC$PTGER4.Group.Celltype,"-",BOC$PTGER2.Group.Celltype)
BOC$PTGER4_PTGER2 = factor(BOC$PTGER4_PTGER2, levels = rev(sort(as.character(unique(BOC$PTGER4_PTGER2)))) )
saveRDS(BOC, "~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.PTGER2Group.240110.rds" )
```

#DE (Bulk)
##PTGER4
```{r}
BOC <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.PTGER2Group.240110.rds")

DE2 = list()
Idents(BOC) = BOC$Identity.Celltype
for(i in levels(BOC$Identity.Celltype)){
      tryCatch({
            print(i)
            DE = list()
            tmp = subset(BOC,idents = i )
            Idents(tmp) = tmp$PTGER4.Group.Celltype
	          DE[["High-Intermediate"]] = FindMarkers(tmp, ident.1 = "High", ident.2 = "Intermediate", logfc.threshold = 0.01, min.pct = 0.01)
            DE[["High-Low"]] = FindMarkers(tmp, ident.1 = "High", ident.2 = "Low", logfc.threshold = 0.01, min.pct = 0.01)
            DE[["High-Undetected"]] = FindMarkers(tmp, ident.1 = "High", ident.2 = "Undetected", logfc.threshold = 0.01, min.pct = 0.01)
            DE[["Intermediate-Low"]] = FindMarkers(tmp, ident.1 = "Intermediate", ident.2 = "Low", logfc.threshold = 0.01, min.pct = 0.01)
            DE[["Intermediate-Undetected"]] = FindMarkers(tmp, ident.1 = "Intermediate", ident.2 = "Undetected", logfc.threshold = 0.01, min.pct = 0.01)
            DE[["Low-Undetected"]] = FindMarkers(tmp, ident.1 = "Low", ident.2 = "Undetected", logfc.threshold = 0.01, min.pct = 0.01)
            tmp = RenameIdents(tmp, 'High' = "Positive",'Intermediate' = "Positive", 'Low' = "Positive",'Undetected' = "Negative")
            DE[["Positive-Negative"]] = FindMarkers(tmp, ident.1 = "Positive",ident.2 = "Negative", logfc.threshold = 0.01, min.pct = 0.01)
            DE2[[i]] = DE
      },error = function(e){print(e)})
}
saveRDS(DE2,"~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Bulk.Celltype.PTGER4.240515.rds")


#DE_bulk.Subtype1.r --------------------------------------------------------------------------------------------------------
DE2 = list()
Idents(BOC) = BOC$Identity.Subtype1
for(i in levels(BOC$Identity.Subtype1)){
      tryCatch({
            print(i)
            DE = list()
            tmp = subset(BOC,idents = i )
            Idents(tmp) = tmp$PTGER4.Group.Subtype1
            DE[["High-Low"]] = FindMarkers(tmp, ident.1 = "High", ident.2 = "Low", logfc.threshold = 0.01, min.pct = 0.01)
            DE[["High-Intermediate"]] = FindMarkers(tmp, ident.1 = "High", ident.2 = "Intermediate", logfc.threshold = 0.01, min.pct = 0.01)
            DE[["High-Undetected"]] = FindMarkers(tmp, ident.1 = "High", ident.2 = "Undetected", logfc.threshold = 0.01, min.pct = 0.01)
            tmp = RenameIdents(tmp, 'High' = "Positive",'Intermediate' = "Positive", 'Low' = "Positive",'Undetected' = "Negative")
            DE[["Positive-Negative"]] = FindMarkers(tmp, ident.1 = "Positive",ident.2 = "Negative", logfc.threshold = 0.01, min.pct = 0.01)
            DE2[[i]] = DE
      },error = function(e){print(e)})
}
saveRDS(DE2,"/~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Bulk.Subtype1.PTGER4.240515.rds")
```


###MultiCondition Comparison 
Both V1 and V2 theoretically give the same result but the V1 code was analyzed under earlier seurat version which yield lower log2fold change compared to Seurat V5. V2 code was done under Seurat V5 so the log2FC is 2-3 times higher.

V1 products; BOCDF$MultipleCom and BOC.DE_Bulk.Celltype.MultipleComparison.240409.rds
V2 products;BOC.DE_Bulk.Celltype.PTGER4.240515.GSEA.rds

####V1
```{r}
tmp = subset(BOC, Identity.Celltype == "T_CD8")
Idents(tmp) = tmp$PTGER4.Group.Celltype
selected.genes = unique(unlist(GeneSet$DE)) %>% intersect(rownames(BOC))

tmp1 = FindMarkers(tmp, ident.1 = "High", ident.2 = "Low", features = selected.genes, min.pct = 0.01, logfc.threshold = 0.01 ) %>% as_tibble(rownames = "Genes")
tmp1$Condition = "High-Low"

tmp2 = FindMarkers(tmp, ident.1 = "High", ident.2 = "Intermediate", features = selected.genes, min.pct = 0.01, logfc.threshold = 0.01 ) %>% as_tibble(rownames = "Genes")
tmp2$Condition = "High-Intermediate"

tmp3 = FindMarkers(tmp, ident.1 = "Intermediate", ident.2 = "Low", features = selected.genes, min.pct = 0.01, logfc.threshold = 0.01 ) %>% as_tibble(rownames = "Genes")
tmp3$Condition = "Intermediate-Low"

tmp4 = FindMarkers(tmp, ident.1 = "High", ident.2 = "Undetected", features = selected.genes, min.pct = 0.01, logfc.threshold = 0.01 ) %>% as_tibble(rownames = "Genes")
tmp4$Condition = "High-Undetected"

tmp5 = FindMarkers(tmp, ident.1 = "Intermediate", ident.2 = "Undetected", features = selected.genes, min.pct = 0.01, logfc.threshold = 0.01 ) %>% as_tibble(rownames = "Genes")
tmp5$Condition = "Intermediate-Undetected"

tmp6 = FindMarkers(tmp, ident.1 = "Low", ident.2 = "Undetected", features = selected.genes, min.pct = 0.01, logfc.threshold = 0.01 ) %>% as_tibble(rownames = "Genes")
tmp6$Condition = "Low-Undetected"

df = rbind(tmp1,tmp2,tmp3,tmp4,tmp5,tmp6)
df$Condition %>% unique()
df$Condition = factor(df$Condition, levels = c("High-Intermediate","High-Low","High-Undetected", "Intermediate-Low","Intermediate-Undetected","Low-Undetected"))
BOCDF$MultipleCom = list("CD8" = df)
#saveRDS(df, "BOC.DE_Bulk.Celltype.MultipleComparison.240409.rds")
```
```{r}
df = BOCDF$MultipleCom$CD8
df1 = GeneGroup(df, GeneSet$DE$ETC.RP)
for(j in 1){
  PDF(paste0("MultiCondition.",i), h = 5 , w = c(7 + (length(unique(df1$Genes))*0.3) ))
  print(
        ggplot(df1, aes(y = Condition, x = Genes  ) ) + theme_classic() + Dot_axis90A +ggtitle(i)+
                  geom_tile(color = "white",aes(fill =avg_log2FC)) +
                  #geom_point(aes(color = avg_log2FC, size = p_val_cat))+scale_size_manual(values = c(2,6,8,10))+
                  #geom_point(aes(color = avg_log2FC),size = 9)+
                  #scale_color_gradientn(colours = Col.HiLow.9Shades, values = scales::rescale(c(min(df1$avg_log2FC),-0.2,0,0.2,max(df1$avg_log2FC)) ), name = "Average\nLog2 FoldChange")+
                  scale_fill_gradientn(colours = Col.HiLow.9Shades,  values = scales::rescale(c(min(df1$avg_log2FC),-0.2,0,0.2,max(df1$avg_log2FC)) ), name = "Average\nLog2 FoldChange")+
                  theme(axis.title = element_blank(), 
                        axis.text.x = element_text(size =16, face = "bold"),
                        axis.text.y = element_text(size =16, face = "bold", color = "black"),
                        strip.text.y.left = element_text(size =18, face = "bold", color = "black",angle = 0),
                        strip.background = element_rect(fill = "white"),
                        strip.placement = "outside",
                        legend.position = "bottom"
                        )
          
      )
  dev.off()
}

df1 = DotPlot(tmp, features = c(GeneSet$DE$MT, GeneSet$DE$MTC, GeneSet$DE$Ribosome, GeneSet$DE$Common))$data

for(j in 1){
  PNG(paste0("MultiConditionScale.",i), h = 5 , w = c(7 + (length(unique(df1$features.plot))*0.3) ))
  print(
    ggplot(df1, aes(y = id, x = features.plot)) + theme_classic() + Dot_axis90A +
      geom_tile(color = "white", aes(fill =avg.exp.scaled)) +
      scale_fill_gradientn(colours = Col.HiLow.9Shades,  values = scales::rescale(c(min(df1$avg.exp.scaled),-0.2,0,0.2,max(df1$avg.exp.scaled)) ))+
      theme(axis.title = element_blank(), 
                        axis.text.x = element_text(size =16, face = "bold"),
                        axis.text.y = element_text(size =16, face = "bold", color = "black"),
                        strip.text.y.left = element_text(size =18, face = "bold", color = "black",angle = 0),
                        strip.background = element_rect(fill = "white"),
                        strip.placement = "outside",
                        legend.position = "bottom"
                        )
  )
  dev.off()
}
```

####V2
```{r}
DE <- readRDS("BOC.DE_Bulk.Celltype.PTGER4.240515.GSEA.rds")
df = data.frame()
for(i in c("High-Low","High-Intermediate","Intermediate-Low","High-Undetected","Intermediate-Undetected","Low-Undetected")){
  tmp1 = DE$T_CD8[[i]] %>% as_tibble(rownames = "Genes") 
  tmp1$Condition = i
  tmp1$Cluster = "T_CD8"
  df = rbind(df,tmp1)
  
}
df$Condition = factor(df$Condition, levels = rev(c("High-Low","High-Intermediate","Intermediate-Low","High-Undetected","Intermediate-Undetected","Low-Undetected")))

```

##PTGER2
```{r}
BOC <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.PTGER2Group.240110.rds")
DE = list()
Idents(BOC) = BOC$Identity.Celltype
for(i in levels(BOC$Identity.Celltype)){
      tryCatch({
            print(i)
            tmp = subset(BOC,idents = i )
            Idents(tmp) = tmp$PTGER2.Group.Celltype
            DE[[i]] = FindMarkers(tmp, ident.1 = "High", ident.2 = "Low", logfc.threshold = 0.01, min.pct = 0.01)
      },error = function(e){print(e)})
}
saveRDS(DE,"~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Bulk.Celltype.PTGER2.240515.rds")
```

##PTGER2-PTGER4
```{r}
BOC <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.PTGER2Group.240110.rds")
DE = list()
Idents(BOC) = BOC$Identity.Celltype
for(i in levels(BOC$Identity.Celltype)){
      tryCatch({
            print(i)
            tmp = subset(BOC,idents = i )
            Idents(tmp) = tmp$PTGER4_PTGER2
            DE[[i]] = FindMarkers(tmp, ident.1 = "High-High", ident.2 = "Low-Low", logfc.threshold = 0.01, min.pct = 0.01)
      },error = function(e){print(e)})
}
#saveRDS(DE,"~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Bulk.PTGER2_PTGER4_Celltype.240516.rds")

DE$T_CD8 %>% arrange(desc(avg_log2FC))
DE$T_CD8[unlist(GeneSet$DE$CD8.Overview3),]

DE$T_CD4_FOXP3[unlist(GeneSet$DE$Chemokine ),]
```


#DE (Per specimen)
##DE PTGER4.Group 
###High - Intermediate
```{r}
BOC <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.PTGER2Group.240110.rds")
DE = list()
for(i in c("T_CD8","TIM") ){
      print(i)
      tmp = subset(BOC, idents = i)
      DE[[i]] = FullDE(tmp, Group = "PTGER4.Group.Celltype", Batch = "Batch.scRNA",A="High",B="Intermediate")
}

df = data.frame()
for(i in names(DE)){
  tmp = names(DE[[i]]) %>% setdiff(c("FoldChange.Table","FoldChange.Sum"))
  for(j in tmp ){
      df1 = DE[[i]][[j]] %>% as_tibble(rownames = "Genes")
      df1$Cluster = i
      df1$Dataset = j
      df = rbind(df,df1)
      }
}
df$Cluster = factor(df$Cluster, levels= names(DE) )
DE$df = df
saveRDS(DE,"~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Specimen.Celltype.PTGER4.High-Intermediate.240515.rds")
```
###High - Low
```{r}
BOC <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.PTGER2Group.240110.rds")
DE = list()
Idents(BOC) = BOC$Identity.Celltype
for(i in levels(Idents(BOC)) ){
      print(i)
      tmp = subset(BOC, idents = i)
      DE[[i]] = FullDE(tmp, Group = "PTGER4.Group.Celltype", Batch = "Batch.scRNA",A = "High", B = "Low")
}
df = data.frame()
for(i in names(DE)){
  tmp = names(DE[[i]]) %>% setdiff(c("FoldChange.Table","FoldChange.Sum"))
  for(j in tmp ){
      df1 = DE[[i]][[j]] %>% as_tibble(rownames = "Genes")
      df1$Cluster = i
      df1$Dataset = j
      df = rbind(df,df1)
      }
}
df$Cluster = factor(df$Cluster, levels= names(DE) )
DE$df = df
saveRDS(DE,"~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Specimen.Celltype.PTGER4.High-Low.240515.rds")

DE = list()
Idents(BOC) = BOC$Identity.Subtype1
for(i in levels(Idents(BOC)) ){
      print(i)
      tmp = subset(BOC, idents = i)
      DE[[i]] = FullDE(tmp, Group = "PTGER4.Group.Subtype1", Batch = "Batch.scRNA")
}
df = data.frame()
for(i in names(DE)){
  tmp = names(DE[[i]]) %>% setdiff(c("FoldChange.Table","FoldChange.Sum"))
  for(j in tmp ){
      df1 = DE[[i]][[j]] %>% as_tibble(rownames = "Genes")
      df1$Cluster = i
      df1$Dataset = j
      df = rbind(df,df1)
      }
}
df$Cluster = factor(df$Cluster, levels= names(DE) )
DE$df = df
saveRDS(DE,"~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Specimen.Subtype1.PTGER4.High-Low.240515.rds")
```

###High - Undetected
```{r}
BOC <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.PTGER2Group.240110.rds")
DE = list()
Idents(BOC) = BOC$Identity.Celltype
for(i in c("T_CD8","TIM") ){
      print(i)
      tmp = subset(BOC, idents = i)
      DE[[i]] = FullDE(tmp, Group = "PTGER4.Group.Celltype", Batch = "Batch.scRNA",A="High",B="Undetected")
}
df = data.frame()
for(i in names(DE)){
  tmp = names(DE[[i]]) %>% setdiff(c("FoldChange.Table","FoldChange.Sum"))
  for(j in tmp ){
      df1 = DE[[i]][[j]] %>% as_tibble(rownames = "Genes")
      df1$Cluster = i
      df1$Dataset = j
      df = rbind(df,df1)
      }
}
df$Cluster = factor(df$Cluster, levels= names(DE) )
DE$df = df
saveRDS(DE,"~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Specimen.Celltype.PTGER4.High-Undetected.240515.rds")
```

###Intermediate-Low
```{r}
BOC = readRDS("//home/siwakorn/Human/RDS/BOC.PTGER4Group.230713.rds")
DE = list()
Idents(BOC) = BOC$Identity.Celltype
for(i in levels(Idents(BOC)) ){
      print(i)
      tmp = subset(BOC, idents = i)
      DE[[i]] = FullDE(tmp, Group = "PTGER4.Group.Celltype", Batch = "Batch.scRNA",A = "Intermediate", B = "Low")
}

df = data.frame()
for(i in names(DE)){
  tmp = names(DE[[i]]) %>% setdiff(c("FoldChange.Table","FoldChange.Sum"))
  for(j in tmp ){
      df1 = DE[[i]][[j]] %>% as_tibble(rownames = "Genes")
      df1$Cluster = i
      df1$Dataset = j
      df = rbind(df,df1)
      }
}
df$Cluster = factor(df$Cluster, levels= names(DE) )
DE$df = df
saveRDS(DE,"~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Specimen.Celltype.PTGER4.Int-Low.240515.rds")
```


##DE PTGER2.Group
```{r}
BOC = readRDS("//home/siwakorn/Human/RDS/BOC.PTGER2Group.240110.rds")
DE = list()
Idents(BOC) = BOC$Identity.Celltype
for(i in levels(Idents(BOC)) ){
      print(i)
      tmp = subset(BOC, idents = i)
      DE[[i]] = FullDE(tmp, Group = "PTGER2.Group.Celltype", Batch = "Batch.scRNA",A = "High", B = "Low")
}

df = data.frame()
for(i in names(DE)){
  tmp = names(DE[[i]]) %>% setdiff(c("FoldChange.Table","FoldChange.Sum"))
  for(j in tmp ){
      df1 = DE[[i]][[j]] %>% as_tibble(rownames = "Genes")
      df1$Cluster = i
      df1$Dataset = j
      df = rbind(df,df1)
      }
}
df$Cluster = factor(df$Cluster, levels= names(DE) )
DE$df = df
saveRDS(DE, "~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Specimen.Celltype.PTGER2.High-Low.240515.rds")
```


##DE PTGER2-PTGER4 Group
```{r}
BOC <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.PTGER2Group.240110.rds")
DE = list()
Idents(BOC) = BOC$Identity.Celltype
for(i in c("T_CD8","T_CD4","TIM") ){
      print(i)
      tmp = subset(BOC, idents = i)
      DE[[i]] = FullDE(tmp, Group = "PTGER4_PTGER2", Batch = "Batch.scRNA",A = "Low-High" , B = "Low-Low")
}
df = data.frame()
for(i in names(DE)){
  tmp = names(DE[[i]]) %>% setdiff(c("FoldChange.Table","FoldChange.Sum"))
  for(j in tmp ){
      df1 = DE[[i]][[j]] %>% as_tibble(rownames = "Genes") 
      df1$Cluster = i
      df1$Dataset = j
      df = rbind(df,df1)
      }
}
DE$df = df
saveRDS(DE, "~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE.PTGER4-2Group.Low-High.Low-Low.240515.rds" )

DE = list()
Idents(BOC) = BOC$Identity.Celltype
for(i in c("T_CD8","T_CD4","TIM") ){
      print(i)
      tmp = subset(BOC, idents = i)
      DE[[i]] = FullDE2(tmp, Group = "PTGER4_PTGER2", Batch = "Batch.scRNA",A = "Undetected-High" , B = "Undetected-Low")
}

df = data.frame()
for(i in names(DE)){
  tmp = names(DE[[i]]) %>% setdiff(c("FoldChange.Table","FoldChange.Sum"))
  for(j in tmp ){
      df1 = DE[[i]][[j]] %>% as_tibble(rownames = "Genes") 
      df1$Cluster = i
      df1$Dataset = j
      df = rbind(df,df1)
      }
}
DE$df = df
saveRDS(DE, "~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE.PTGER4-2Group.Un-High.Un-Low.240515.rds" )


DE = list()
Idents(BOC) = BOC$Identity.Celltype
for(i in c("T_CD8","T_CD4","TIM") ){
      print(i)
      tmp = subset(BOC, idents = i)
      DE[[i]] = FullDE2(tmp, Group = "PTGER4_PTGER2", Batch = "Batch.scRNA",A = "High-Low" , B = "Low-Low")
}

df = data.frame()
for(i in names(DE)){
  tmp = names(DE[[i]]) %>% setdiff(c("FoldChange.Table","FoldChange.Sum"))
  for(j in tmp ){
      df1 = DE[[i]][[j]] %>% as_tibble(rownames = "Genes") 
      df1$Cluster = i
      df1$Dataset = j
      df = rbind(df,df1)
      }
}
DE$df = df
saveRDS(DE, "~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE.PTGER4-2Group.High-Low.Low-Low.240515.rds" )


DE = list()
Idents(BOC) = BOC$Identity.Celltype
for(i in c("T_CD8","T_CD4","TIM") ){
      print(i)
      tmp = subset(BOC, idents = i)
      DE[[i]] = FullDE2(tmp, Group = "PTGER4_PTGER2", Batch = "Batch.scRNA",A = "High-Undetected" , B = "Low-Undetected")
}

df = data.frame()
for(i in names(DE)){
  tmp = names(DE[[i]]) %>% setdiff(c("FoldChange.Table","FoldChange.Sum"))
  for(j in tmp ){
      df1 = DE[[i]][[j]] %>% as_tibble(rownames = "Genes") 
      df1$Cluster = i
      df1$Dataset = j
      df = rbind(df,df1)
      }
}
DE$df = df
saveRDS(DE, "~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE.PTGER4-2Group.High-Un.Low-Un.240515.rds" )

DE = list()
Idents(BOC) = BOC$Identity.Celltype
for(i in c("T_CD8","T_CD4","T_CD4_FOXP3") ){
      print(i)
      tmp = subset(BOC, idents = i)
      DE[[i]] = FullDE2(tmp, Group = "PTGER4_PTGER2", Batch = "Batch.scRNA",A = "High-High" , B = "Low-Low")
}

df = data.frame()
for(i in names(DE)){
  tmp = names(DE[[i]]) %>% setdiff(c("FoldChange.Table","FoldChange.Sum"))
  for(j in tmp ){
      df1 = DE[[i]][[j]] %>% as_tibble(rownames = "Genes") 
      df1$Cluster = i
      df1$Dataset = j
      df = rbind(df,df1)
      }
}
DE$df = df
saveRDS(DE, "~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE.PTGER4-2Group.High-High.Low-Low.240515.rds" )
```


#GSEA
##Bulk (Celltype)
```{r}
#DE <- readRDS("~/RStudioProject/scRNA/BOC.DE_Bulk.Celltype.230713.rds")
DE <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Bulk.Celltype.230713.rds")
df = data.frame()
for(i in names(DE)){
  tmp = names(DE[[i]]) 
  for(j in tmp ){
      df1 = DE[[i]][[j]] %>% as_tibble(rownames = "Genes")
      df1$Cluster = i
      df1$Condition = j
      df = rbind(df,df1)
      }
}

###GSEA
GSEA = list()
GSEA_H = list()
for(i in names(DE)[1:8] ){
  tmp = DE[[i]][["High-Low"]] %>% as_tibble(rownames = "Genes" ) %>% dplyr::select(Genes, avg_log2FC) %>% arrange(avg_log2FC)
  GSEA[[i]] <- fgsea(fgsea_sets, stats = deframe(tmp), nperm = 50000) %>% as_tibble() %>% arrange(desc(NES) )
  GSEA_H[[i]] <- fgsea(fgsea_sets_H, stats = deframe(tmp), nperm = 50000) %>% as_tibble() %>% arrange(desc(NES) )
}
DE$GSEA = GSEA
DE$GSEA_H = GSEA_H

###GSEA p < 0.05
GSEA = list()
GSEA_H = list()
for(i in names(DE)[1:8] ){
  tmp = DE[[i]][["High-Low"]] %>% as_tibble(rownames = "Genes" ) %>% filter(p_val < 0.05) %>% dplyr::select(Genes, avg_log2FC) %>% arrange(avg_log2FC)
  GSEA[[i]] <- fgsea(fgsea_sets, stats = deframe(tmp), nperm = 50000) %>% as_tibble() %>% arrange(desc(NES) )
  GSEA_H[[i]] <- fgsea(fgsea_sets_H, stats = deframe(tmp), nperm = 50000) %>% as_tibble() %>% arrange(desc(NES) )
}
DE$GSEA_p05 = GSEA
DE$GSEA_H_p05 = GSEA_H

###GSEA Adj pval < 0.05
GSEA = list()
GSEA_H = list()
for(i in names(DE)[1:8] ){
  tmp = DE[[i]][["High-Low"]] %>% as_tibble(rownames = "Genes" ) %>% filter(p_val_adj < 0.05) %>% dplyr::select(Genes, avg_log2FC) %>% arrange(avg_log2FC)
  GSEA[[i]] <- fgsea(fgsea_sets, stats = deframe(tmp), nperm = 50000) %>% as_tibble() %>% arrange(desc(NES) )
  GSEA_H[[i]] <- fgsea(fgsea_sets_H, stats = deframe(tmp), nperm = 50000) %>% as_tibble() %>% arrange(desc(NES) )
}
DE$GSEA_padj05 = GSEA
DE$GSEA_H_padj05 = GSEA_H

###GSEA Multilevel
GSEA = list()
GSEA_H = list()
for(i in names(DE)[1:8] ){
  tmp = DE[[i]][["High-Low"]] %>% as_tibble(rownames = "Genes" ) %>% dplyr::select(Genes, avg_log2FC) %>% arrange(avg_log2FC)
  GSEA[[i]] <- fgseaMultilevel(fgsea_sets, stats = deframe(tmp) ) %>% arrange(desc(NES) )
  GSEA_H[[i]] <- fgseaMultilevel(fgsea_sets_H, stats = deframe(tmp) ) %>% arrange(desc(NES) )
}
DE$GSEA2 = GSEA
DE$GSEA_H2 = GSEA_H

DE$df.HighLow = filter(df, Condition == "High-Low")
DE$df.HighInt = filter(df, Condition == "High-Intermediate")
DE$df.HighUndetected= filter(df, Condition == "High-Undetected")
DE$df.PosNeg = filter(df, Condition == "Positive-Negative")

#saveRDS(DE,"~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Bulk.Celltype.Converted.GSEA_nperm50K.2300901.rds")
DE = readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Bulk.Celltype.Converted.GSEA_nperm50K.2300901.rds")
```

##Per Specimen Celltype
```{r}
DE2 <- readRDS("BOC.DE_Specimen.Celltype.230713.rds")
df = data.frame()
for(i in names(DE2)){
  tmp = names(DE2[[i]]) %>% setdiff(c("FoldChange.Table","FoldChange.Sum"))
  for(j in tmp ){
      df1 = DE2[[i]][[j]] %>% as_tibble(rownames = "Genes")
      df1$Cluster = i
      df1$Dataset = j
      df = rbind(df,df1)
      }
}
df$Cluster = factor(df$Cluster, levels= names(DE2) )
DE2$df = df
GSEA = list()
for(j in c("T_CD8","TIM")){
  tmp1 = setdiff(names(DE2[[j]] ),c("FoldChange.Table", "FoldChange.Sum"))
  df = data.frame()
  for(i in tmp1){
    tmp = DE2[[j]][[i]] %>% as_tibble(rownames = "Genes" ) %>% dplyr::select(Genes, avg_log2FC) %>% arrange(avg_log2FC)
    ranks<- deframe(tmp)
    fgseaRes1 <- fgsea(fgsea_sets, stats = ranks, nperm = 1000)
    df1 = fgseaRes1 %>% arrange(NES)
    df1$Dataset = i
    df = rbind(df,df1)
  }
  df$Cluster = j
  GSEA[[j]] = df
}
DE2$GSEA = GSEA
saveRDS(DE2,"~/RStudioProject/scRNA/BOC.DE_Specimen.Celltype.230713.converted.rds")
```

##Per Specimen Celltype4
```{r}
DE2 <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Specimen.Celltype4.230713.rds")
df = data.frame()
for(i in names(DE2)){
  tmp = names(DE2[[i]]) %>% setdiff(c("FoldChange.Table","FoldChange.Sum"))
  for(j in tmp ){
      df1 = DE2[[i]][[j]] %>% as_tibble(rownames = "Genes")
      df1$Cluster = i
      df1$Dataset = j
      df = rbind(df,df1)
      }
}
df$Cluster = factor(df$Cluster, levels= names(DE2) )
DE2$df = df

GSEA = list()
for(j in c("T_CD8","TIM")){
  tmp1 = setdiff(names(DE2[[j]] ),c("FoldChange.Table", "FoldChange.Sum"))
  df = data.frame()
  for(i in tmp1){
    tmp = DE2[[j]][[i]] %>% as_tibble(rownames = "Genes" ) %>% dplyr::select(Genes, avg_log2FC) %>% arrange(avg_log2FC)
    ranks<- deframe(tmp)
    fgseaRes1 <- fgsea(fgsea_sets, stats = ranks, nperm = 1000)
    df1 = fgseaRes1 %>% arrange(NES)
    df1$Dataset = i
    df = rbind(df,df1)
  }
  df$Cluster = j
  GSEA[[j]] = df
}
DE2$GSEA = GSEA
#saveRDS(DE2,"~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Specimen.Celltype4.230713.converted.rds")
DE2 = readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Specimen.Celltype4.230713.converted.rds")
```

##Per Specimen Subtype1
```{r}
DE2 <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Specimen.Subtype1.230713.rds")

df = data.frame()
for(i in names(DE2)[c(3:5,11)]){
  tmp = names(DE2[[i]]) %>% setdiff(c("FoldChange.Table","FoldChange.Sum"))
  for(j in tmp ){
      df1 = DE2[[i]][[j]] %>% as_tibble(rownames = "Genes")
      df1$Cluster = i
      df1$Dataset = j
      df = rbind(df,df1)
      }
}
df$Cluster = factor(df$Cluster, levels = names(DE2))
DE2$df.CD8 = df

df = data.frame()
for(i in names(DE2)[19:22] ){
  tmp = names(DE2[[i]]) %>% setdiff(c("FoldChange.Table","FoldChange.Sum"))
  for(j in tmp ){
      df1 = DE2[[i]][[j]] %>% as_tibble(rownames = "Genes")
      df1$Cluster = i
      df1$Dataset = j
      df = rbind(df,df1)
      }
}
df$Cluster = factor(df$Cluster, levels = names(DE2))
DE2$df.TIM = df
#saveRDS(DE2,"~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Specimen.Subtype1.230713.converted.rds")
```

#=====LLC1=======
##Celltype2
```{r}
DE <- readRDS("~/RStudioProject/LLC/DE.LLC.Celltype2.230714.rds")
#DE = readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/LLC/DE.LLC.Celltype3.230719.rds")
selected.cluster = setdiff(names(DE$Day6),"Doublet")
df = data.frame()
for(t in  c("Day1.5","Day6")){
  for(i in  selected.cluster ){
    tmp = DE[[t]][[i]] %>% as_tibble(rownames = "Genes")
    if(!is.null(tmp)){
      tmp$Dataset = t
      tmp$Cluster = i
      tmp$Condition = "EP2iEP4i-Control"
      df = rbind(df,tmp) }
  }
}
df$Dataset = factor(df$Dataset,levels = c("Day1.5","Day6"))
df$Cluster = factor(df$Cluster, levels = selected.cluster )
LLCDF$df.DE.Celltype2 = df

```

##Celltype3
```{r}
DE = readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/LLC/DE.LLC.Celltype3.230719.rds")
selected.cluster = setdiff(names(DE$Day6),"Doublet")
df = data.frame()
for(t in  c("Day1.5","Day6")){
  for(i in  selected.cluster ){
    tmp = DE[[t]][[i]] %>% as_tibble(rownames = "Genes")
    if(!is.null(tmp)){
      tmp$Dataset = t
      tmp$Cluster = i
      tmp$Condition = "EP2iEP4i-Control"
      df = rbind(df,tmp) }
  }
}
df$Dataset = factor(df$Dataset,levels = c("Day1.5","Day6"))
df$Cluster = factor(df$Cluster, levels = selected.cluster )
LLCDF$df.DE.Celltype3 = df

```



```{r}
BOC <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.PTGER2Group.240110.rds")
#DE <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/BOC.DE_Bulk.Celltype.PTGER4.240515.GSEA.rds")
DE <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/Archive_240515_PublicationFile/BOC.DE_Bulk.Celltype.Converted.GSEA_nperm50K.2300901.rds")
DE <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/Archive_240515_PublicationFile/BOC.DE_Bulk.Subtype1.230713.rds")

DE <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/Archive_240515_PublicationFile/BOC.DE_Bulk.Celltype.240409.rds") #Seurat 5
DE <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/Archive_240515_PublicationFile/BOC.DE_Bulk.PTGER2_Celltype.240411.rds") #Seurat 5

DE <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/Archive_240515_PublicationFile/BOC.DE_Specimen.Celltype.230713.converted.rds")
DE <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/Archive_240515_PublicationFile/BOC.DE_Specimen.Subtype1.230713.converted.rds")

df <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA/Archive_240515_PublicationFile/BOC.DE_Bulk.Celltype.MultipleComparison.240409.rds")

DE <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA.Public/Pelka/DE.AllGenes.Pelka.PTGER4.Group2.Celltype4.240408.rds")


DE <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/LLC/used/DE.LLC.Celltype3.pval.nperm50k.230903.rds")

DE <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/PublicBulkRNA/GSE47189.Xue.Schultze.Macrophage/DE.Xue.230909.rds")

M23DF <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/CD8_RNAseq/CD8_RANseq_M23DF.rds")

DE <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/PublicBulkRNA/Schultze.CD4/DE.CD4.Schultze-Meyer.230908.rds")
Beyer <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/PublicBulkRNA/Schultze.CD4/Beyer.230829.rds")

Bona <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA.Public/Bonavita/Bonavita.QC.230714.rds")
DE <- readRDS("~/Library/CloudStorage/OneDrive-Personal/Documents_Onedrive/RStudioProject/scRNA.Public/Bonavita/DE.Bonavita.Celltype2.230714.rds")
```
